[#20]feat: contextApi를 사용한 데이터 전달
