+++ src/components/Keyword/KeywordSubmit/styles.js
@@ -1,0 +1,18 @@
+import styled from 'styled-components';
+
+export const EditButton = styled.button`
+  width: 80px;
+  height: 32px;
+  background: #222;
+  color: #fff;
+  position: absolute;
+  top: 834px;
+  left: 1018px;
+  font-size: 14px;
+  font-weight: 500;
+  text-align: center;
+`;
+
+export const CancelButton = styled(EditButton)`
+  left: 1142px;
+`;
