+++ src/pages/MainPage.js
@@ -1,19 +1,8 @@
 import React from 'react';
-import SideNavbar from 'components/SideNavbar';
-import AddKeyword from 'components/Keyword/ModifyKeyword';
-import styled from 'styled-components';
-
-const Container = styled.div`
-  display: flex;
-`;
+import KeywordModifyPage from './KeywordModifyPage';
 
 const MainPage = () => {
-  return (
-    <Container>
-      <SideNavbar />
-      <AddKeyword />
-    </Container>
-  );
+  return <KeywordModifyPage />;
 };
 
 export default MainPage;
